move component to folder, get rid of example code and unnecessary functions
